lerna create @firestore-simple/admin and fix types
