+++ packages/web/README.md
@@ -1,0 +1,11 @@
+# `@firestore-simple/web`
+
+> TODO: description
+
+## Usage
+
+```
+const web = require('@firestore-simple/web');
+
+// TODO: DEMONSTRATE API
+```
