+++ packages/web/__tests__/basic.test.ts
@@ -1,5 +1,5 @@
 import { firestore } from 'firebase'
-import { FirestoreSimpleWeb } from '../../src/'
+import { FirestoreSimpleWeb } from '../src/'
 import { WebFirestoreTestUtil } from './util'
 
 const util = new WebFirestoreTestUtil()
