+++ packages/web/__tests__/encode.test.ts
@@ -1,4 +1,4 @@
-import { FirestoreSimpleWeb } from '../../src'
+import { FirestoreSimpleWeb } from '../src'
 import { WebFirestoreTestUtil } from './util'
 const util = new WebFirestoreTestUtil()
 const webFirestore = util.webFirestore
