+++ packages/web/src/index.ts
@@ -1,0 +1,4 @@
+// Web SDK
+export { FirestoreSimpleWeb } from './web'
+export type { WebCollection } from './collection'
+export type { WebEncodable, WebDecodable } from './types'
