--- src/index.ts
@@ -1,13 +1,0 @@
-// Ailias name for campatible before v6.0.0
-export { FirestoreSimpleAdmin as FirestoreSimple } from './admin/admin'
-export type { AdminEncodable as Encodable, AdminDecodable as Decodable } from './admin/types'
-
-// Admin SDK
-export { FirestoreSimpleAdmin } from './admin/admin'
-export type { AdminCollection } from './admin/collection'
-export type { AdminEncodable, AdminDecodable } from './admin/types'
-
-// Web SDK
-export { FirestoreSimpleWeb } from './web/web'
-export type { WebCollection } from './web/collection'
-export type { WebEncodable, WebDecodable } from './web/types'
